requirements.txt
pandas>=1.5.0
requests>=2.28.0
reportlab>=4.0.0
matplotlib>=3.6.0
gspread>=6.0.0
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
openai>=1.0.0